backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
ortools==9.11.4210
pydantic==2.9.2
python-multipart==0.0.12
websockets==13.1
numpy==2.1.3
python-dateutil==2.9.0
python-dotenv==1.0.1
aiofiles==24.1.0
loguru==0.7.2